requirements.txt
streamlit==1.22.0
streamlit-drawable-canvas==0.9.2
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
psycopg2-binary==2.9.6
numpy==1.24.2